requirements.txt
streamlit
numpy
tensorflow
Pillow
streamlit-option-menu
streamlit-webrtc